Sync 2015-01-18

+++ media-gfx/fbida/fbida-2.09-r4.ebuild
@@ -1,6 +1,6 @@
-# Copyright 1999-2014 Gentoo Foundation
+# Copyright 1999-2015 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/media-gfx/fbida/fbida-2.09-r4.ebuild,v 1.1 2014/11/27 01:53:32 jer Exp $
+# $Header: /var/cvsroot/gentoo-x86/media-gfx/fbida/fbida-2.09-r4.ebuild,v 1.9 2015/01/02 12:37:35 ago Exp $
 
 EAPI=5
 inherit eutils toolchain-funcs
@@ -14,7 +14,7 @@ SRC_URI="
 "
 LICENSE="GPL-2 IJG"
 SLOT="0"
-KEYWORDS="~alpha ~amd64 ~arm ~hppa ~ppc ~ppc64 ~sh ~sparc ~x86"
+KEYWORDS="~alpha amd64 arm hppa ppc ppc64 ~sh sparc x86"
 IUSE="curl fbcon +gif lirc pdf +png scanner +tiff X +webp"
 REQUIRED_USE="
 	pdf? ( tiff )
@@ -111,7 +111,7 @@ src_install() {
 
 	dodoc README
 
-	if ! use pdf; then
+	if use fbcon && ! use pdf; then
 		rm "${D}"/usr/bin/fbgs "${D}"/usr/share/man/man1/fbgs.1 || die
 	fi
